Add typing improvements and new package dependencies
- Introduced type annotations across various modules, enhancing type safety and clarity in the codebase.
- Added `types-pillow` package to `pyproject.toml` for improved type checking with Pillow.
- Updated `poetry.lock` to include the new package and ensure compatibility with Python versions.
- Refined function signatures and variable types in multiple files to align with MyPy standards, contributing to overall code quality.
- Enhanced documentation to reflect the new typing standards and package dependencies.

+++ src/giflab/meta.py
@@ -223,4 +223,4 @@ def calculate_entropy(image: Image.Image) -> float:
     # Ensure non-negative result (handle floating point precision issues)
     entropy = max(0.0, entropy)
 
-    return round(entropy, 3)
+    return float(round(entropy, 3))
